sf: remove event listener when it was fired.
when GFAR is enable, sf dispsync resource is keeping
alive to monitor janks. When jank happens, previous
event listener is removed, and with new listener added at
once. when mHasFired is set to true, we need remove this
listener and with new one added. It can make the
mLastCallbackTime update correctly.

Change-Id: I2c1879f82656c03972d7506726d6898211693fff
CRs-Fixed: 2586618
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>